backend/requirements.txt
# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-SocketIO==5.3.5
python-socketio==5.10.0
eventlet==0.33.3

# Database
pymongo==4.6.1
dnspython==2.4.2

# Machine Learning
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
joblib==1.3.2

# Computer Vision & OCR
opencv-python==4.8.1.78
pytesseract==0.3.10
Pillow==10.1.0

# Face Recognition (Optional - use OpenCV alternative)
# face-recognition==1.3.0
# dlib==19.24.2

# PDF Generation
reportlab==4.0.7

# Utilities
python-dotenv==1.0.0
Werkzeug==3.0.1